+++ scripts/lib/constants.js
@@ -125,6 +125,7 @@ const actorCompendiumPacks = () => [{label: '', value: ''}, ...game.packs.filter
 const itemCompendiumPacks = () => [{label: '', value: ''}, ...game.packs.filter(i => i.documentName === 'Item').map(i => ({label: i.metadata.label, value: i.metadata.id}))];
 const abilityOptions = () => Object.values(CONFIG.DND5E.abilities).map(i => ({label: i.label, value: i.abbreviation}));
 const healingTypeOptions = () => Object.entries(CONFIG.DND5E.healingTypes).map(i => ({label: i[1].label, value: i[0]}));
+const diceSizeOptions = [4, 6, 8, 10, 12, 20].map(i => ({label: 'd' + i, value: 'd' + i}));
 const overTimeOptions = [
     {
         key: 'turn',
@@ -473,5 +474,6 @@ export let constants = {
     healingTypeOptions,
     legacyPacks,
     modernPacks,
-    immuneEffectData
+    immuneEffectData,
+    diceSizeOptions
 };