+++ scripts/macros/2024/generic/proneOnFail.js
@@ -1,0 +1,7 @@
+import {proneOnFail as proneOnFailLegacy} from '../../generic/proneOnFail.js';
+export let proneOnFail = {
+    name: proneOnFailLegacy.name,
+    verson: proneOnFailLegacy.version,
+    rules: 'modern',
+    midi: proneOnFailLegacy.midi
+};